KVM: arm64: Pass a 64bit function-id in the SMC handlers
Make the SMC handlers accept a 64bit value for the function-id to keep
it uniform with the rest of the code and prevent a u64 -> u32 -> u64
conversion as it currently happens when we handle PSCI.

Message-Id: <20260401123201.389906-1-sebastianene@google.com>

+++ arch/arm64/kvm/hyp/nvhe/ffa.c
@@ -862,7 +862,7 @@ out_unlock:
 	hyp_spin_unlock(&host_buffers.lock);
 }
 
-bool kvm_host_ffa_handler(struct kvm_cpu_context *host_ctxt, u32 func_id)
+bool kvm_host_ffa_handler(struct kvm_cpu_context *host_ctxt, u64 func_id)
 {
 	struct arm_smccc_1_2_regs res;
